ms_rutinas_django/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
djangorestframework-simplejwt==5.3.1
cryptography==42.0.5
django-cors-headers
django-environ==0.11.2
kafka-python==2.0.2